CLN: move fibre_cell_region_swap to shg_pl_trans module
